requirements.txt
Django==1.9
django-report-builder==3.4.4
djangorestframework==3.5.3
et-xmlfile==1.0.1
jdcal==1.3
MySQL-python==1.2.5
openpyxl==2.3.5
python-dateutil==2.6.0
six==1.10.0
django-admin-tools
django-filter
django-allauth
django-rest-auth
django-localflavor